refactor(website): to re-organize components
