+++ website/src/components/index.js
@@ -1,0 +1,4 @@
+export { default as BurgerIcon } from './burger-icon';
+export { default as CodeBlock } from './code-block';
+export { default as SEO } from './seo';
+export { default as TextHighlight } from './text-highlight';
