+++ website/src/components/text-highlight.js
@@ -1,8 +1,7 @@
 import styled from '@emotion/styled';
 import * as colors from '../colors';
 
-// eslint-disable-next-line import/prefer-default-export
-export const TextHighlight = styled.span`
+const TextHighlight = styled.span`
   background-image: linear-gradient(
     ${colors.light.primarySoft},
     ${colors.light.primarySoft}
@@ -11,3 +10,5 @@ export const TextHighlight = styled.span`
   background-size: 100% 0.2em;
   background-position: 0 85%;
 `;
+
+export default TextHighlight;
