+++ website/src/layouts/index.js
@@ -1,0 +1,2 @@
+export { default as LayoutContent } from './content';
+export { default as LayoutMarketing } from './marketing';
