+++ website/src/layouts/internals/layout-container.js
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const LayoutContainer = styled.div`
+  height: 100vh;
+  display: grid;
+  grid-template-rows: 50px 1fr;
+  grid-template-columns: auto 1fr auto;
+`;
+
+export default LayoutContainer;
