+++ website/src/layouts/internals/navbar.js
@@ -4,8 +4,8 @@ import { useStaticQuery, graphql, Link } from 'gatsby';
 import { css, ClassNames } from '@emotion/core';
 import styled from '@emotion/styled';
 import { Spacings, customProperties } from '@commercetools-frontend/ui-kit';
-import { TextHighlight } from './styled';
-import * as colors from '../colors';
+import { TextHighlight } from '../../components';
+import * as colors from '../../colors';
 
 const NavbarLinkTitle = styled.div`
   font-size: 1.2rem;
